requirements.txt
asgiref==3.2.3
click==7.1.1
Django==3.0.4
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.4
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.0
